requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.7
python-dotenv==1.0.0
requests==2.31.0
Flask-Migrate==4.0.7
Flask-Limiter==3.8.0
pytest==8.3.3

